[refactor] : 체험 상세 디자인 수정 및 스켈레톤 적용
